worker_service/requirements.txt
pika
psycopg2-binary
pydantic
pydantic-settings
